requirements.txt
arrow==0.12.0
Django==1.11.7
django-filter==1.1.0
django-mathfilters==0.4.0
django-tagging==0.4.6
django-timezone-field==2.0
django-widget-tweaks==1.4.1
Pillow==4.3.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-decouple==3.1
social-auth-app-django==2.0.0
twilio==6.9.0
virtualenv==15.1.0